Stop using an ACM policy to replicate a secret in multiplenamespace
Let's just create a helm template that loops over the namespaces that
are needed and let the externalsecrets be automatically in sync with the
s3 password.

This has the benefit of simplifying ACM and having Argo less confused
about object that were created by ACM and not by Argo itself.

Let's also remove the previous policy