Notebook tweaks
In RHOAI 2.22 there are automatic mounts for the CAs. So for now
let's just drop our initcontainer that pulled the cluster's CA
and just git clone ignoring the self-signed cert from gitea.

In a second step we will need to figure out what the kosher way
to do this in 2.22 and later, is.